@@ -3,44 +3,43 @@
     "build": "gulp"
   },
   "devDependencies": {
-    "autoprefixer-core": "^5.1.11",
-    "babel-core": "^4.3.0",
-    "babel-loader": "^5.0.0",
-    "browser-sync": "^2.5.3",
-    "cmi": "^2.0.0",
-    "gulp": "^3.8.11",
+    "autoprefixer-core": "^5.2.0",
+    "babel-core": "^5.5.8",
+    "babel-loader": "^5.1.4",
+    "browser-sync": "^2.7.12",
+    "cmi": "^2.0.4",
+    "gulp": "^3.9.0",
     "gulp-browserify": "^0.5.1",
     "gulp-concat": "^2.5.2",
-    "gulp-csslint": "^0.1.5",
-    "gulp-jshint": "^1.10.0",
-    "gulp-minify-css": "^1.0.0",
-    "gulp-postcss": "^5.1.3",
-    "gulp-react": "^3.0.1",
+    "gulp-minify-css": "^1.1.6",
+    "gulp-postcss": "^5.1.8",
     "gulp-rename": "^1.2.2",
-    "gulp-replace": "^0.5.3",
-    "gulp-size": "^1.2.1",
-    "gulp-sourcemaps": "^1.5.1",
-    "gulp-uglify": "^1.1.0",
-    "gulp-util": "^3.0.4",
+    "gulp-uglify": "^1.2.0",
+    "gulp-util": "^3.0.5",
     "gulp-watch": "^4.2.4",
-    "jsx-loader": "^0.13.1",
-    "postcss": "^4.1.6",
-    "postcss-calc": "^4.0.1",
-    "postcss-color-function": "^1.2.0",
-    "postcss-color-rgba-fallback": "^1.1.1",
-    "postcss-for": "^0.2.0",
+    "jsx-loader": "^0.13.2",
+    "merge": "^1.2.0",
+    "postcss": "^4.1.11",
+    "postcss-assets": "^2.1.2",
+    "postcss-calc": "^4.1.0",
+    "postcss-color-function": "^1.3.0",
+    "postcss-color-gray": "^2.0.0",
+    "postcss-color-rgba-fallback": "^1.2.0",
+    "postcss-each": "^0.2.0",
+    "postcss-for": "^1.0.1",
     "postcss-image-set": "^1.0.0",
-    "postcss-import": "^5.2.2",
-    "postcss-merge-rules": "^1.2.2",
-    "postcss-mixins": "^0.1.1",
-    "postcss-nested": "^0.2.2",
+    "postcss-import": "^6.0.0",
+    "postcss-merge-rules": "^1.3.0",
+    "postcss-mixins": "^0.3.0",
+    "postcss-nested": "^0.3.2",
     "postcss-opacity": "^2.0.0",
-    "postcss-pxtorem": "^2.0.1",
+    "postcss-position": "^0.2.0",
+    "postcss-pxtorem": "^2.1.0",
     "postcss-simple-extend": "^0.3.1",
-    "postcss-simple-vars": "^0.2.4",
-    "postcss-size": "^0.1.0",
-    "postcss-will-change": "^0.1.0",
-    "through2": "^0.6.5",
-    "webpack": "^1.8.4"
+    "postcss-simple-vars": "^0.3.0",
+    "postcss-size": "^0.2.0",
+    "postcss-will-change": "^0.2.0",
+    "require-dir": "^0.3.0",
+    "webpack": "^1.9.11"
   }
 }
